requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfminer.six==20221105
python-docx==1.1.0
